refactoring institution info of details page
