+++ src/components/InstitutionPage/DetailsPage/index.js
@@ -1,64 +1,16 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { Fade, Paper, Typography, withStyles, Button } from '@material-ui/core';
+import { Fade, withStyles } from '@material-ui/core';
 import { connect } from 'react-redux';
-import styled from 'styled-components';
-import { ActionsContainer } from '../../utils/SharedComponents';
 import GradesGrid from './GradesGrid';
 import CreateEntryCodeDialog from '../EntryCode/CreateEntryCodeDialog';
 import DisplayCodeDialog from '../EntryCode/DisplayCodeDialog';
 import EmployeeList from './EmployeeList';
 import { requestGraphql } from '../../utils/HTTPClient';
 import { getUsersFromInstitutionByRole } from '../../../queryGenerators/userQueries';
-
-const photoDimension = '140px';
-
-const InstitutionInfos = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  flex-wrap: wrap;
-`;
-
-const contentContainerBreakpoints = theme => ({
-  [theme.breakpoints.up('sm')]: {
-    marginLeft: theme.spacing.unit * 3,
-  },
-});
+import InstitutionInfos from './InstitutionInfo';
 
 const styles = theme => ({
-  button: {
-    marginBottom: theme.spacing.unit,
-    marginLeft: theme.spacing.unit,
-    marginTop: theme.spacing.unit,
-  },
-  institutionInfos: {
-    marginBottom: theme.spacing.unit * 3,
-    marginTop: theme.spacing.unit * 3,
-    ...contentContainerBreakpoints(theme),
-  },
-  photo: {
-    height: photoDimension,
-    marginBottom: theme.spacing.unit * 3,
-    marginTop: theme.spacing.unit * 3,
-    width: photoDimension,
-    ...contentContainerBreakpoints(theme),
-  },
-  detailsPageContentContainer: {
-    display: 'flex',
-    [theme.breakpoints.down('xs')]: {
-      alignItems: 'center',
-      flexDirection: 'column',
-    },
-  },
-  detailsPagePaper: {
-    borderRadius: 0,
-    display: 'flex',
-    flexDirection: 'column',
-    marginTop: theme.spacing.unit * 3,
-    padding: theme.spacing.unit,
-    width: `calc(100% - ${theme.spacing.unit * 2}px)`,
-  },
   detailsPageRoot: {
     display: 'flex',
     flexDirection: 'column',
@@ -75,9 +27,6 @@ const styles = theme => ({
       width: '60%',
     },
   },
-  actionsDetailsButton: {
-    marginLeft: theme.spacing.unit,
-  },
 });
 
 class DetailsPage extends Component {
@@ -90,10 +39,6 @@ class DetailsPage extends Component {
       teachers: [],
       admins: [],
     };
-
-    this.handleCreateEntryOpenToggle = this.handleCreateEntryOpenToggle.bind(this);
-    this.handleDisplayCodeOpenToggle = this.handleDisplayCodeOpenToggle.bind(this);
-    this.fetchData = this.fetchData.bind(this);
   }
 
   componentDidMount() {
@@ -110,7 +55,7 @@ class DetailsPage extends Component {
     }
   }
 
-  fetchData() {
+  fetchData = () => {
     const teacherInput = {
       institutionId: this.props.institution.id,
       role: 'TEACHER',
@@ -141,12 +86,12 @@ class DetailsPage extends Component {
     .catch();
   }
 
-  handleDisplayCodeOpenToggle() {
+  handleDisplayCodeOpenToggle = () => {
     this.setState(({ displayCodeOpen }) =>
       ({ displayCodeOpen: !displayCodeOpen }));
   }
 
-  handleCreateEntryOpenToggle(currentGeneratedCode = null) {
+  handleCreateEntryOpenToggle = (currentGeneratedCode = null) => {
     if (currentGeneratedCode && typeof currentGeneratedCode === 'string') {
       this.setState({
         entryCodeCreateOpen: false,
@@ -180,49 +125,10 @@ class DetailsPage extends Component {
             open={entryCodeCreateOpen}
             onClose={this.handleCreateEntryOpenToggle}
           />
-          <Paper className={classes.detailsPagePaper}>
-            <div className={classes.detailsPageContentContainer}>
-              <img
-                alt="Institution"
-                className={classes.photo}
-                src="https://s3.amazonaws.com/tinycards/image/f8bda7d1c863b4f42adf2d1e5d72ff14" />
-              <InstitutionInfos className={classes.institutionInfos}>
-                <Typography
-                  variant="title"
-                  gutterBottom
-                >
-                  {institution.name}
-                </Typography>
-                <Typography
-                  variant="subheading"
-                  gutterBottom
-                >
-                  Email: {institution.email}
-                </Typography>
-                <Typography
-                  variant="subheading"
-                  gutterBottom
-                >
-                  Telefone: {institution.phone}
-                </Typography>
-                <Typography
-                  gutterBottom
-                  variant="subheading"
-                >
-                  Endereço: {institution.address}
-                </Typography>
-              </InstitutionInfos>
-            </div>
-            <ActionsContainer>
-              <Button
-                className={classes.actionsDetailsButton}
-                color="primary"
-                onClick={this.handleCreateEntryOpenToggle}
-              >
-                Gerar Código de Vínculo
-              </Button>
-            </ActionsContainer>
-          </Paper>
+          <InstitutionInfos
+            institution={institution}
+            onHandleCreateEntryOpenToggle={this.handleCreateEntryOpenToggle}
+          />
           <GradesGrid />
           <EmployeeList employees={teachers} headTitle="Professores" />
           <EmployeeList employees={admins} headTitle="Administradores" />
